Allow recursion until stack overflow is hit I have jsonnet code ressembling the `assert_recursion_allowed.jsonnet` file This file evaluates fine in C jsonnet and go jsonnet but fails here because of `infinite recursion detected`
The fix I propose here is to now fail when recursion is detected but rather keep going until stack overflow is reached (as is done in other jsonnet interpreters)
I also modified the test for issue #23. This issue is still fixed IMO because it doesn't panic. Instead, it prints out the full stack, which is what other jsonnet interpreters do, and it is a meaningful error message as well

Note: I am a complete Rust noob. My fix may be terrible. Let me know if it is!